server-fastapi/pyproject.toml
[project]
name = "agents-raw"
version = "0.1.0"
description = "FastAPI server for agentic system"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.86.0",
    "fastapi>=0.135.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.42.0",
]
